+++ src/utils/iamBindings/typeToKeyGeneratorMap.ts
@@ -71,8 +71,8 @@ import {
   ORGANIZATION_ENTITY_TYPE,
   FOLDER_ENTITY_TYPE,
   PROJECT_ENTITY_TYPE,
-} from '../../steps/resource-manager';
-import { API_SERVICE_ENTITY_TYPE } from '../../steps/service-usage';
+} from '../../steps/resource-manager/constants';
+import { API_SERVICE_ENTITY_TYPE } from '../../steps/service-usage/constants';
 import {
   ENTITY_TYPE_SPANNER_INSTANCE,
   ENTITY_TYPE_SPANNER_INSTANCE_DATABASE,